折扣功能开发完成

+++ group-buy-market-infrastructure/src/main/java/com/binghu/gbm/infrastructure/dao/po/GroupBuyDiscount.java
@@ -1,5 +1,6 @@
 package com.binghu.gbm.infrastructure.dao.po;
 
+import com.binghu.gbm.domain.activity.model.valobj.DiscountTypeEnum;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
@@ -35,7 +36,7 @@ public class GroupBuyDiscount {
     /**
      * 折扣类型（0:base、1:tag）
      */
-    private Byte discountType;
+    private Integer discountType;
 
     /**
      * 营销优惠计划（ZJ:直减、MJ:满减、N元购）
